+++ src/components/text/text.js
@@ -6,8 +6,8 @@ export default class ManagedText extends ManagedInput {
   constructor(props) {
     super(props)
     this.type = this.props.manager('type')
-    this.autoComplete = props.manager('autoComplete')
-    this.match = props.manager('match')
+    this.autoComplete = this.props.manager('autoComplete')
+    this.match = this.props.manager('match')
     this.containerClass = this.classNames.container || 'managed-input text'
     this.onKeyDownHandler = handlers.onKeyDownHandler.bind(this)
     this.onChangeHandler = handlers.onChangeHandler.bind(this)
@@ -36,6 +36,7 @@ export default class ManagedText extends ManagedInput {
         onKeyDown={this.onKeyDownHandler}
         onBlur={this.onBlurHandler}
         onChange={this.onChangeHandler} />
+        {this.props.manager('children')}
         <div className={this.errorContainerClass + this.state.decorator}>
           {this.state.error}
         </div>